requirements.txt
# Data Collection
requests==2.31.0
pandas==2.0.3
python-dateutil==2.8.2
feedparser==6.0.10

# Kafka
kafka-python==2.0.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Vector DB
pymilvus==2.3.4

# Embedding & ML
sentence-transformers==2.3.1
transformers==4.38.0
torch==2.2.0
bitsandbytes==0.42.0
accelerate==0.27.0

# Processing
pathway==0.2.1

# RAG
langchain==0.2.0
langchain-community==0.2.19

# API
fastapi==0.109.0
uvicorn==0.25.0

# Utilities
schedule==1.2.0
numpy==1.24.4
tabulate==0.9.0